Try to change the webpack configuration for Angular (with Injection)
- not works, error message :
Uncaught Error: Can't resolve all parameters for t: (?).
